move logger to __init__
